+++ ellie_engine/tokenizer/src/processors/types/decimal_processor.rs
@@ -35,7 +35,7 @@ impl crate::processors::Processor for decimal_type::DecimalTypeCollector {
             } else if letter_char.to_string().parse::<i8>().is_ok() {
                 self.point += &letter_char.to_string();
                 self.data.raw += &letter_char.to_string();
-                if let Ok(nm) = self.data.raw.parse::<f64>() {
+                if let Ok(nm) = self.data.raw.parse::<f32>() {
                     self.data.value = DecimalTypeEnum::Float(nm);
                     self.complete = true;
                 } else {
@@ -51,7 +51,7 @@ impl crate::processors::Processor for decimal_type::DecimalTypeCollector {
                 self.data.pos.range_end = cursor;
             } else if letter_char == 'f' {
                 self.data.pos.range_end = cursor;
-                if let Ok(nm) = self.data.raw.parse::<f64>() {
+                if let Ok(nm) = self.data.raw.parse::<f32>() {
                     self.data.value = DecimalTypeEnum::Float(nm);
                     self.data.is_double = false;
                     self.complete = true;
@@ -67,7 +67,7 @@ impl crate::processors::Processor for decimal_type::DecimalTypeCollector {
                 }
             } else if letter_char == 'd' {
                 self.data.pos.range_end = cursor;
-                if let Ok(nm) = self.data.raw.parse::<f32>() {
+                if let Ok(nm) = self.data.raw.parse::<f64>() {
                     self.data.value = DecimalTypeEnum::Double(nm);
                     self.complete = false;
                 } else {
